samples/onnx_accelerator_sample2/source/backend/requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.js and run "npx projen".
aws-cdk-lib>=2.70.0, <3.0.0
aws-cdk.aws-apigatewayv2-alpha
aws-cdk.aws-apigatewayv2-authorizers-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha
boto3==1.26.72
cdk-nag==2.22.28
constructs>=10.0.5, <11.0.0
